Rename Node to HotblocksServer
