+++ Dockerfile
@@ -11,14 +11,14 @@ ADD Cargo.lock .
 ADD crates crates
 
 
-FROM builder AS node-builder
-RUN cargo build -p sqd-node-example --release
+FROM builder AS hotblocks-builder
+RUN cargo build -p sqd-hotblocks-example --release
 
 
-FROM rust AS node-example
+FROM rust AS hotblocks-example
 WORKDIR /app
-COPY --from=node-builder /app/target/release/sqd-node-example .
-RUN mkdir -p /etc/sqd/node && echo "{}" > /etc/sqd/node/datasets.json
-VOLUME /var/sqd/node
-ENTRYPOINT ["/app/sqd-node-example", "--db", "/var/sqd/node/db", "--datasets", "/etc/sqd/node/datasets.json"]
+COPY --from=hotblocks-builder /app/target/release/sqd-hotblocks-example .
+RUN mkdir -p /etc/sqd/hotblocks && echo "{}" > /etc/sqd/hotblocks/datasets.json
+VOLUME /var/sqd/hotblocks
+ENTRYPOINT ["/app/sqd-hotblocks-example", "--db", "/var/sqd/hotblocks/db", "--datasets", "/etc/sqd/hotblocks/datasets.json"]
 EXPOSE 3000
