+++ crates/hotblocks/src/hotblocks/mod.rs
@@ -1,9 +1,8 @@
-mod node;
-mod node_builder;
 mod query_executor;
 mod query_response;
+mod server;
+mod server_builder;
 
-
-pub use node::*;
-pub use node_builder::*;
 pub use query_response::QueryResponse;
+pub use server::*;
+pub use server_builder::*;
