+++ crates/hotblocks/src/lib.rs
@@ -1,11 +1,11 @@
 #![allow(unused)]
 pub mod error;
 mod ingest;
-mod node;
+mod hotblocks;
 mod query;
 mod types;
 
 
-pub use node::*;
+pub use hotblocks::*;
 pub use types::{DBRef, DatasetKind, RetentionStrategy};
 pub use sqd_query::Query;
